+++ src/main/java/tragicneko/tragicmc/util/WorldHelper.java
@@ -258,9 +258,9 @@ public class WorldHelper {
 		float f = 1.0F;
 		float f1 = ent.prevRotationPitch + (ent.rotationPitch - ent.prevRotationPitch) * f;
 		float f2 = ent.prevRotationYaw + (ent.rotationYaw - ent.prevRotationYaw) * f;
-		double d0 = ent.prevPosX + (ent.posX - ent.prevPosX) * (double)f;
-		double d1 = ent.prevPosY + (ent.posY - ent.prevPosY) * (double)f + (double)(ent.getEyeHeight());
-		double d2 = ent.prevPosZ + (ent.posZ - ent.prevPosZ) * (double)f;
+		double d0 = ent.prevPosX + (ent.posX - ent.prevPosX) * f;
+		double d1 = ent.prevPosY + (ent.posY - ent.prevPosY) * f + (ent.getEyeHeight());
+		double d2 = ent.prevPosZ + (ent.posZ - ent.prevPosZ) * f;
 		Vec3 vec3 = Vec3.createVectorHelper(d0, d1, d2);
 		float f3 = MathHelper.cos(-f2 * 0.017453292F - (float)Math.PI);
 		float f4 = MathHelper.sin(-f2 * 0.017453292F - (float)Math.PI);
@@ -274,7 +274,7 @@ public class WorldHelper {
 		{
 			d3 = ((EntityPlayerMP)ent).theItemInWorldManager.getBlockReachDistance() + (d3 - 4.0D);
 		}
-		Vec3 vec31 = vec3.addVector((double)f7 * d3, (double)f6 * d3, (double)f8 * d3);
+		Vec3 vec31 = vec3.addVector(f7 * d3, f6 * d3, f8 * d3);
 
 		return ent.worldObj.func_147447_a(vec3, vec31, true, false, true);
 	}
